mongo_stream_proj/requirements.txt
boto3==1.18.6
pandas==1.3.1
pymongo==3.12.0
python-dotenv==0.19.1
